move uiManager code into own folder
